+++ portal-impl/src/com/liferay/portal/ccpp/PortalProfileFactory.java
@@ -16,7 +16,12 @@ package com.liferay.portal.ccpp;
 
 import com.sun.ccpp.ProfileFactoryImpl;
 
+import java.util.Set;
+
+import javax.ccpp.Attribute;
+import javax.ccpp.Component;
 import javax.ccpp.Profile;
+import javax.ccpp.ProfileDescription;
 import javax.ccpp.ProfileFactory;
 import javax.ccpp.ValidationMode;
 
@@ -48,6 +53,33 @@ public class PortalProfileFactory {
 		return profile;
 	}
 
-	private static final Profile _profile = new EmptyProfile();
+	private static final Profile _profile = new Profile() {
+
+		@Override
+		public Attribute getAttribute(String name) {
+			return null;
+		}
+
+		@Override
+		public Set<Attribute> getAttributes() {
+			return null;
+		}
+
+		@Override
+		public Component getComponent(String localtype) {
+			return null;
+		}
+
+		@Override
+		public Set<Component> getComponents() {
+			return null;
+		}
+
+		@Override
+		public ProfileDescription getDescription() {
+			return null;
+		}
+
+	};
 
 }
